Submit ready file
I tested this code and it seems to be working. It goes into the data and pulls the species list, by common name where possible for the given genus. It also asks if you want to pull images and if so, it proceeds to copy a raw image of the web.